Final Edition
Provide 2 Functionality to achieve usability of this project
1. query command : ?#struct
Displaying the structure of the current ASTNode. Like where this node
belongs to.
2. query command: ?#tree
Displaying the current structure of the whole AST. This is useful to
check whether a block statement is complete and for the user to view the
current AST structure. In this query, if the current node is in an
incomplete block statement of the program, the whole block statement
will not be displayed in the tree.

Final Feature: The program will not only generate the C code syntax
after finish parsing the input
but generate the text form of  AST of the program